encapsulate `ensure_dll` better
