+++ rust/bin-shared/benches/tunnel.rs
@@ -40,9 +40,6 @@ mod platform {
     use tun::Tun as _;
 
     pub(crate) async fn perf() -> Result<()> {
-        // Install wintun so the test can run
-        let _wintun_path = firezone_bin_shared::windows::ensure_dll().unwrap();
-
         const MTU: usize = 1_280;
         const NUM_REQUESTS: u64 = 1_000;
         const REQ_CODE: u8 = 42;
